aggregate-datasets: add LinkModelAbstractShould; Add LinkModelFactory interface; refactor ModelFactory and ModelAbstractShould; Apply changes in DataSetDataElementLinkModelShould
