+++ core/src/main/java/org/hisp/dhis/android/core/common/LinkModelFactory.java
@@ -28,6 +28,8 @@
 
 package org.hisp.dhis.android.core.common;
 
-public interface ModelFactory<M extends Model, P> extends LinkModelFactory<M> {
-    public M fromPojo(P p);
+import android.database.Cursor;
+
+public interface LinkModelFactory<M extends Model> {
+    public M fromCursor(Cursor cursor);
 }
